+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Extend link validity to have 'From ... to ...'
     - Define new 'validity-period-created', 'validity-period-left-merged', 'validity-period-right-merged' and 'validity-period-removed' events
     - Improve free validity text
+    - Fix validity panel and setup buttons title
 
 ### 1.3.0